requirements.txt
requests
pdfplumber
python-dotenv
msal
pandas
odfpy
numpy
sys
os
glob
ezodf
lxml
pymupdf

